debugging env

+++ my_modules/mapping.py
@@ -132,15 +132,15 @@ def set_neighborhood_around_obstacle(x, y):
     global environment, FUZZ_FACTOR, ROOM_HEIGHT_CM, ROOM_WIDTH_CM
     round_x, round_y = round(x), round(y)
     
-    x_s = np.arange(round_x - FUZZ_FACTOR, round_x + FUZZ_FACTOR + 1)
-    y_s = np.arange(round_y - FUZZ_FACTOR, round_y + FUZZ_FACTOR + 1)
+    x_s = np.arange(round_x - FUZZ_FACTOR, round_x + FUZZ_FACTOR + 1, dtype=np.int32)
+    y_s = np.arange(round_y - FUZZ_FACTOR, round_y + FUZZ_FACTOR + 1, dtype=np.int32)
     
     print("before trim")
     print(x_s)
     print(y_s)
     
     x_s = x_s[x_s >= 0 and x_s < ROOM_WIDTH_CM]
-    y_s = y_s[y_s >=0 and y_s < ROOM_HEIGHT_CM]
+    y_s = y_s[y_s >= 0 and y_s < ROOM_HEIGHT_CM]
     
     print("after trim")
     print(x_s)
